+++ lib/cache/constructor_test.go
@@ -22,10 +22,12 @@ package cache
 
 import (
 	"os"
+	"strings"
 	"testing"
 
 	"github.com/Jeffail/benthos/lib/log"
 	"github.com/Jeffail/benthos/lib/metrics"
+	"github.com/Jeffail/benthos/lib/types"
 	yaml "gopkg.in/yaml.v3"
 )
 
@@ -44,6 +46,32 @@ func TestConstructorBadType(t *testing.T) {
 
 	if _, err := New(conf, nil, log.New(os.Stdout, logConfig), metrics.DudType{}); err == nil {
 		t.Error("Expected error, received nil for invalid type")
+	}
+}
+
+func TestConstructorBlockType(t *testing.T) {
+	Constructors["footype"] = TypeSpec{
+		constructor: func(
+			conf Config,
+			mgr types.Manager,
+			log log.Modular,
+			stats metrics.Type,
+		) (types.Cache, error) {
+			return nil, nil
+		},
+	}
+
+	conf := NewConfig()
+	conf.Type = "footype"
+
+	Block("footype", "because testing")
+
+	_, err := New(conf, nil, log.Noop(), metrics.Noop())
+	if err == nil {
+		t.Fatal("Expected error, received nil for blocked type")
+	}
+	if !strings.Contains(err.Error(), "because testing") {
+		t.Errorf("Unexpected error: %v", err)
 	}
 }